Adding CFO Run Record support and feature cleanup in preparation for Global Run 2
