+++ otsdaq-mu2e/CFOandDTCCore/CFOandDTCCoreVInterface.h
@@ -84,8 +84,8 @@ class CFOandDTCCoreVInterface : public FEVInterface
 
 
   public: 
-	void 								SoftReset						(__ARGS__);
-	void 								HardReset						(__ARGS__);
+	void 								SoftReset							(__ARGS__);
+	void 								HardReset							(__ARGS__);
 
 	void								GetFirmwareVersion					(__ARGS__);
 	void 								ResetPCIe							(__ARGS__);
