+++ test/CFOFrontEndInterface_TestMain.cc
@@ -16,24 +16,23 @@
 #include "otsdaq/TableCore/MakeTable.h"
 
 // #include "otsdaq-components/FEInterfaces/FEWOtsUDPFSSRInterface.h"
-#include "otsdaq-mu2e/FEInterfaces/DTCFrontEndInterface.h"
+#include "otsdaq-mu2e/FEInterfaces/CFOFrontEndInterface.h"
 
 using namespace ots;
 
 int main(int argc, char* argv[])
 try
 {
-   	__COUT_INFO__ << "DTCFrontEndInterface Test main()";
+   	__COUT_INFO__ << "CFOFrontEndInterface Test main()";
 
 	__COUTV__(argc);
 	for(int i=0;i<argc;++i)
 	{
 		__COUT_INFO__ << "arg[" << i << "] = " << argv[i] << __E__;
 	}
-	if(argc < 3)
+	if(argc != 3)
 	{
-		__COUT_ERR__ << "\n\n\tUsage = Need at least 2 arguments: DTCFrontEndInterface_TestMain <deviceIndex> <numberOfEventWindowMarkers>\n\n" << __E__;
-		__COUT_ERR__ << "\n\n\t\t 3+ aruments will apply ROC emulator data generation size.\n\n" << __E__;
+		__COUT_ERR__ << "\n\n\tUsage = Need 2 arguments: CFOFrontEndInterface_TestMain <deviceIndex> <numberOfEventWindowMarkers>\n\n" << __E__;		
 		return 0;
 	}
 
@@ -90,7 +89,7 @@ try
 	// 	    .getValue<std::string>();
 	// __COUTV__(name);
 
-    DTCFrontEndInterface dtc(feUID_,
+    CFOFrontEndInterface cfo(feUID_,
         cfgMgr.getNode(ConfigurationManager::XDAQ_CONTEXT_TABLE_NAME),
         theConfigurationPath_);
 
@@ -99,84 +98,70 @@ try
 	{
 		__COUT_INFO__ << "Attempting Reset and Buffer Release ONLY!" << __E__;
 
-		dtc.thisDTC_->DisableCFOEmulation();
-		dtc.thisDTC_->SoftReset();
-		dtc.thisDTC_->ReleaseAllBuffers(DTC_DMA_Engine_DAQ);
+		cfo.getCFOandDTCRegisters()->SetJitterAttenuatorSelect(1 /* select RJ45 */, false /* alsoResetJA */);
+		sleep(1);
+    	__COUT_INFO__ << "JA Status = " << cfo.getCFOandDTCRegisters()->FormatJitterAttenuatorCSR() << __E__;
+		cfo.thisCFO_->SoftReset();
+		cfo.thisCFO_->ReleaseAllBuffers(CFO_DMA_Engine_DAQ);
 		__COUT_INFO__ << "Reset and ReleaseAllBuffers called!" << __E__;
 		return 0;
 	}
 	if(numberOfEventWindowMarkers == uint32_t(-2))
 	{
 		__COUT_INFO__ << "Attempting Buffer Release ONLY!" << __E__;
-		dtc.thisDTC_->ReleaseAllBuffers(DTC_DMA_Engine_DAQ);
+		cfo.thisCFO_->ReleaseAllBuffers(CFO_DMA_Engine_DAQ);
 		__COUT_INFO__ << "ReleaseAllBuffers called!" << __E__;
 		return 0;
 	}
 	if(numberOfEventWindowMarkers == uint32_t(-3))
 	{
 		__COUT_INFO__ << "Setting to 32KB Max DMA Transfer size!" << __E__;
-		dtc.thisDTC_->SetTriggerDMATransferLength(0x8000);
-		__COUT_INFO__ << "DTC DMA sizes = " << dtc.thisDTC_->FormatDMATransferLength() << __E__;
+		cfo.thisCFO_->SetTriggerDMATransferLength(0x8000);
+		__COUT_INFO__ << "CFO DMA sizes = " << cfo.thisCFO_->FormatDMATransferLength() << __E__;
 		return 0;
 	}
 	if(numberOfEventWindowMarkers == uint32_t(-4))
 	{
 		__COUT_INFO__ << "Setting to 64KB Max DMA Transfer size!" << __E__;
-		dtc.thisDTC_->SetTriggerDMATransferLength(0xFFF8);
-		__COUT_INFO__ << "DTC DMA sizes = " << dtc.thisDTC_->FormatDMATransferLength() << __E__;
+		cfo.thisCFO_->SetTriggerDMATransferLength(0xFFF8);
+		__COUT_INFO__ << "CFO DMA sizes = " << cfo.thisCFO_->FormatDMATransferLength() << __E__;
 		return 0;
 	}
 
-    __COUT_INFO__ << "DTC version = " << dtc.thisDTC_->ReadDesignDate() << __E__;
-	__COUT_INFO__ << "DTC DMA sizes = " << dtc.thisDTC_->FormatDMATransferLength() << __E__;
+    __COUT_INFO__ << "CFO version = " << cfo.thisCFO_->ReadDesignDate() << __E__;
+	__COUT_INFO__ << "CFO DMA sizes = " << cfo.thisCFO_->FormatDMATransferLength() << __E__;
 
-	//setup ROCs
-	std::string reply;
-	for(int i=3;i<argc;++i)
+
+	cfo.getCFOandDTCRegisters()->SetJitterAttenuatorSelect(1 /* select RJ45 */, false /* alsoResetJA */);
+	for(int i=0;i<10;++i) //wait for JA to lock before reading
 	{
-		int sz = atoi(argv[i]);
-		__COUT_INFO__ << "ROC #" << i-3 << " size arg[" << i << "] = " << sz << __E__;
-
-		if(sz == -1) //disabled
-			reply = dtc.SetupROCs(
-				DTCLib::DTC_Link_ID(i-3), //]DTCLib::DTC_Link_ID rocLinkIndex,
-				0, 1, 1, //bool rocRxTxEnable, bool rocTimingEnable, bool rocEmulationEnable,
-				DTCLib::DTC_ROC_Emulation_Type(0 /* 0: Internal, 1: Fiber-Loopback, 2: External */),// DTCLib::DTC_ROC_Emulation_Type rocEmulationType,
-				0// uint32_t size
-			);
-		else
-			reply = dtc.SetupROCs(
-				DTCLib::DTC_Link_ID(i-3), //]DTCLib::DTC_Link_ID rocLinkIndex,
-				1, 1, 1, //bool rocRxTxEnable, bool rocTimingEnable, bool rocEmulationEnable,
-				DTCLib::DTC_ROC_Emulation_Type(0 /* 0: Internal, 1: Fiber-Loopback, 2: External */),// DTCLib::DTC_ROC_Emulation_Type rocEmulationType,
-				atoi(argv[i])// uint32_t size
-			);
+		if(cfo.getCFOandDTCRegisters()->ReadJitterAttenuatorLocked())
+			break;
+		sleep(1);
 	}
-	__COUT_INFO__ << "ROC Setup:\n" << reply << __E__;
-    
-    dtc.SetCFOEmulatorFixedWidthEmulation(
+    __COUT_INFO__ << "JA Status = " << cfo.getCFOandDTCRegisters()->FormatJitterAttenuatorCSR() << __E__;
+
+    cfo.CompileSetAndLaunchTemplateFixedWidthRunPlan(
         1, //bool enable, 
         1, //bool useDetachedBufferTest,
         "0x44 clocks", //std::string eventDuration, 
         numberOfEventWindowMarkers, //uint32_t numberOfEventWindowMarkers, 
-        0, //uint64_t initialEventWindowTag,
-        1, //uint64_t eventWindowMode, 
+        100, //uint64_t initialEventWindowTag,
+        0x333, //uint64_t eventWindowMode, 
         0, //bool enableClockMarkers, 
-        1, //bool enableAutogenDRP, 
         0, //bool saveBinaryDataToFile,
         0, //bool saveSubeventHeadersToDataFile,
         0  //bool doNotResetCounters )
     );
-	// dtc.SetCFOEmulatorOnOffSpillEmulation(
+	// cfo.CompileSetAndLaunchTemplateSuperCycleRunPlan(
 	// 	1, //bool enable,
 	// 	1, //bool useDetachedBufferTest, 
 	// 	numberOfEventWindowMarkers, //uint32_t numberOfSuperCycles, 
 	// 	0, //uint64_t initialEventWindowTag,
-    //     0, //bool enableClockMarkers, 
-    //     1, //bool enableAutogenDRP, 
-    //     0, //bool saveBinaryDataToFile,
-    //     0, //bool saveSubeventHeadersToDataFile,
-    //     0  //bool doNotResetCounters )
+    //  0, //bool enableClockMarkers, 
+    //  0, //bool saveBinaryDataToFile,
+    //  0, //bool saveSubeventHeadersToDataFile,
+    //  0  //bool doNotResetCounters )
 	// ); numberOfEventWindowMarkers *= 245000; //set event cout expectation for check below (235K on-spill + 10K off-spill per cycle)
 
 	int i=0;
@@ -189,31 +174,31 @@ try
 		{
 			std::cout << "time(0) = " << time(0) << '\n' << std::flush;
 			__COUT_INFO__ << "\n" << 
-				DTCFrontEndInterface::getDetachedBufferTestStatus(dtc.bufferTestThreadStruct_) << __E__;	
+				CFOFrontEndInterface::getDetachedBufferTestStatus(cfo.bufferTestThreadStruct_) << __E__;	
 		}
 
-		if((i > 5 && !dtc.bufferTestThreadStruct_->running_) || 
-			DTCFrontEndInterface::getDetachedBufferTestReceivedCount(dtc.bufferTestThreadStruct_) >= numberOfEventWindowMarkers - 1)// start mutex scope
+		if((i > 5 && !cfo.bufferTestThreadStruct_->running_) || 
+			CFOFrontEndInterface::getDetachedBufferTestReceivedCount(cfo.bufferTestThreadStruct_) >= numberOfEventWindowMarkers - 1)// start mutex scope
 		{
-			__COUT_INFO__ << "Iteration exit #" << i << " - thread running = " << dtc.bufferTestThreadStruct_->running_ << "\n" << 
-				DTCFrontEndInterface::getDetachedBufferTestStatus(dtc.bufferTestThreadStruct_) << __E__;	
+			__COUT_INFO__ << "Iteration exit #" << i << " - thread running = " << cfo.bufferTestThreadStruct_->running_ << "\n" << 
+				CFOFrontEndInterface::getDetachedBufferTestStatus(cfo.bufferTestThreadStruct_) << __E__;	
 
 			sleep(1);
-			std::lock_guard<std::mutex> lock(dtc.bufferTestThreadStruct_->lock_);
-			dtc.bufferTestThreadStruct_->exitThread_ = true;
+			std::lock_guard<std::mutex> lock(cfo.bufferTestThreadStruct_->lock_);
+			cfo.bufferTestThreadStruct_->exitThread_ = true;
 
-			if(DTCFrontEndInterface::getDetachedBufferTestReceivedCount(dtc.bufferTestThreadStruct_) != numberOfEventWindowMarkers - 1)
+			if(CFOFrontEndInterface::getDetachedBufferTestReceivedCount(cfo.bufferTestThreadStruct_) < numberOfEventWindowMarkers - 1)
 				dumpSpy = true;			
 			break;
 		}
 		++i;
 	} //end main loop
 	sleep(1);
-	if(dtc.bufferTestThreadStruct_->running_)
+	if(cfo.bufferTestThreadStruct_->running_)
 		sleep(1); //give 1 more second for thread
 
 	if(dumpSpy)
-		dtc.getDevice()->spy(DTC_DMA_Engine_DAQ, 3 /* for once */ | 8 /* for wide view */ | 16 /* for stack trace */);
+		cfo.getDevice()->spy(CFO_DMA_Engine_DAQ, 3 /* for once */ | 8 /* for wide view */ | 16 /* for stack trace */);
 
 	__COUT_INFO__ << "Thread and main exited!" << __E__;
 	return 0;
