+++ test/DTCFrontEndInterface_ExtCFOMain.cc
@@ -23,7 +23,7 @@ using namespace ots;
 int main(int argc, char* argv[])
 try
 {
-   	__COUT_INFO__ << "DTCFrontEndInterface Test main()";
+   	__COUT_INFO__ << "DTCFrontEndInterface External CFO main()";
 
 	__COUTV__(argc);
 	for(int i=0;i<argc;++i)
@@ -32,7 +32,7 @@ try
 	}
 	if(argc < 3)
 	{
-		__COUT_ERR__ << "\n\n\tUsage = Need at least 2 arguments: DTCFrontEndInterface_TestMain <deviceIndex> <numberOfEventWindowMarkers>\n\n" << __E__;
+		__COUT_ERR__ << "\n\n\tUsage = Need at least 2 arguments: DTCFrontEndInterface_ExtCFOMain <deviceIndex> <numberOfEventWindowMarkers>\n\n" << __E__;
 		__COUT_ERR__ << "\n\n\t\t 3+ aruments will apply ROC emulator data generation size.\n\n" << __E__;
 		return 0;
 	}
@@ -94,12 +94,15 @@ try
         cfgMgr.getNode(ConfigurationManager::XDAQ_CONTEXT_TABLE_NAME),
         theConfigurationPath_);
 
-
+	
 	if(numberOfEventWindowMarkers == uint32_t(-1))
 	{
 		__COUT_INFO__ << "Attempting Reset and Buffer Release ONLY!" << __E__;
 
 		dtc.thisDTC_->DisableCFOEmulation();
+		dtc.getCFOandDTCRegisters()->SetJitterAttenuatorSelect(1 /* select RJ45 */, false /* alsoResetJA */);
+		sleep(1);
+    	__COUT_INFO__ << "JA Status = " << dtc.getCFOandDTCRegisters()->FormatJitterAttenuatorCSR() << __E__;
 		dtc.thisDTC_->SoftReset();
 		dtc.thisDTC_->ReleaseAllBuffers(DTC_DMA_Engine_DAQ);
 		__COUT_INFO__ << "Reset and ReleaseAllBuffers called!" << __E__;
@@ -130,6 +133,12 @@ try
     __COUT_INFO__ << "DTC version = " << dtc.thisDTC_->ReadDesignDate() << __E__;
 	__COUT_INFO__ << "DTC DMA sizes = " << dtc.thisDTC_->FormatDMATransferLength() << __E__;
 
+	dtc.SetupCFOInterface(
+		0, //int forceCFOedge, 
+		false, //bool useCFOemulator, 
+		true, //bool cfoRxTxEnable, 
+		true); //bool enableAutogenDRP);
+
 	//setup ROCs
 	std::string reply;
 	for(int i=3;i<argc;++i)
@@ -153,32 +162,13 @@ try
 			);
 	}
 	__COUT_INFO__ << "ROC Setup:\n" << reply << __E__;
-    
-    dtc.SetCFOEmulatorFixedWidthEmulation(
-        1, //bool enable, 
-        1, //bool useDetachedBufferTest,
-        "0x44 clocks", //std::string eventDuration, 
-        numberOfEventWindowMarkers, //uint32_t numberOfEventWindowMarkers, 
-        0, //uint64_t initialEventWindowTag,
-        1, //uint64_t eventWindowMode, 
-        0, //bool enableClockMarkers, 
-        1, //bool enableAutogenDRP, 
-        0, //bool saveBinaryDataToFile,
-        0, //bool saveSubeventHeadersToDataFile,
-        0  //bool doNotResetCounters )
-    );
-	// dtc.SetCFOEmulatorOnOffSpillEmulation(
-	// 	1, //bool enable,
-	// 	1, //bool useDetachedBufferTest, 
-	// 	numberOfEventWindowMarkers, //uint32_t numberOfSuperCycles, 
-	// 	0, //uint64_t initialEventWindowTag,
-    //     0, //bool enableClockMarkers, 
-    //     1, //bool enableAutogenDRP, 
-    //     0, //bool saveBinaryDataToFile,
-    //     0, //bool saveSubeventHeadersToDataFile,
-    //     0  //bool doNotResetCounters )
-	// ); numberOfEventWindowMarkers *= 245000; //set event cout expectation for check below (235K on-spill + 10K off-spill per cycle)
 
+	dtc.thisDTC_->SoftReset(); //to reset event window tag starting point handling
+	dtc.initDetachedBufferTest(100,//initialEventWindowTag,
+			false, // saveBinaryDataToFile, 
+			false, //saveSubeventHeadersToDataFile, 
+			false); //doNotResetCounters);
+    
 	int i=0;
 	bool dumpSpy = false;
 	while(1)
